Games | Combine commands to create test data (#184)
Combine commands to create test data
